Add files via upload
Title: Weather App using API
Description: This repo contains the code for a weather app that displays the current weather conditions for a given location. The app is built with HTML, JavaScript, Bootstrap, and an API.
Files:
index.html: The main HTML file that contains the markup for the app.
script.js: The JavaScript file that contains the code for the app's functionality.
Dependencies:
Bootstrap
API : https://openweathermap.org/
The API of your choice
Instructions:
Clone the repo.
Install the dependencies.
Update the API key in the api.js file.
Run the app.
The app should now be running and displaying the current weather conditions for the location you specified.

Here are some additional things you could do to improve the app:

Add support for multiple locations.
weather forecast.
current weather conditions.
I hope this helps!